Fix commit 1eec350
Commit 1eec350 - "Fix checking delay after timer expires before running"
was not correct. This should resolve the errors.

+++ lib/process.c
@@ -40,7 +40,7 @@
 
 static unsigned cur_rt_priority;
 static unsigned max_rt_priority;
-unsigned min_auto_priority_delay;
+long min_auto_priority_delay;
 
 #if HAVE_DECL_RLIMIT_RTTIME == 1
 static unsigned cur_rlimit_rttime;
@@ -111,7 +111,7 @@ reset_process_priority(void)
    variable length buffer" warning */
 RELAX_STACK_PROTECTOR_START
 void
-set_process_priorities(int realtime_priority, int max_realtime_priority, unsigned min_delay,
+set_process_priorities(int realtime_priority, int max_realtime_priority, long min_delay,
 #if HAVE_DECL_RLIMIT_RTTIME == 1
 		       int rlimit_rt,
 #endif
